farmtech_final/requirements.txt
# FarmTech ERP - Dependências Essenciais

# Core Database e Backend
SQLAlchemy>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
PyYAML>=6.0.0

# Machine Learning e Análise
scikit-learn>=1.4.0
statsmodels>=0.14.0

# Interface Web (Streamlit)
streamlit>=1.33.0
plotly>=5.0.0

# Utilitários
requests>=2.31.0 